refactor: update import paths for CanDoListWrapper component
