+++ frontend/components/dashboard/can-do-list/can-do-list-wrapper.tsx
@@ -4,7 +4,7 @@ import dynamic from 'next/dynamic';
 
 // Dynamically import the CanDoListPage component with no SSR
 const CanDoListPage = dynamic(
-  () => import('../../app/dashboard/can-do-list/page'),
+  () => import('@/app/dashboard/can-do-list/page'),
   { ssr: false }
 );
